Make a viable python 3 compatible version Change abstract parser to a simpler base adapter (closes [#8](https://github.com/everyonesdesign/enru-python/issues/8))
