+++ bin/enru
@@ -4,7 +4,7 @@ import pydoc
 import click
 
 from enru.enru import Enru
-from enru.parsers.yandex import YandexParser
+from enru.adapters.yandex import YandexAdapter
 
 
 @click.command()
@@ -15,9 +15,9 @@ from enru.parsers.yandex import YandexParser
               help='Show word usage examples')
 def main(word, show_examples):
     """Simple dictionaries translation parser"""
-    parser = YandexParser()
+    parser = YandexAdapter()
     enru = Enru(parser)
-    content = enru.run(word, show_examples)
+    content = enru.run(word)
     pydoc.pipepager(content, cmd='less -R')
 
 if __name__ == "__main__":
