--- enru/parsers/abstract.py
@@ -1,13 +1,0 @@
-from abc import ABCMeta, abstractmethod
-
-
-class AbstractParser:
-    __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def get_url(self):
-        pass
-
-    @abstractmethod
-    def get_content(self):
-        pass
